requirements.txt
ffmpeg_python==0.2.0
gTTS==2.5.4
moviepy==2.1.2
mutagen==1.47.0
openai_whisper==20240930
pysrt==1.1.2
Requests==2.32.3
selenium==4.32.0
selenium_stealth==1.0.6
